+++ packages/core/src/client/index.ts
@@ -8,7 +8,7 @@ export {
 } from '../graphql-request-client';
 export { DefaultRetryStrategy } from '../retries';
 export { RetryStrategy, PageInfo, FetchOptions } from '../models';
-export { getEdgeProxyContentUrl, getEdgeProxyFormsUrl } from './graphql-edge-proxy';
+export { getEdgeProxyContentUrl, getEdgeProxyFormsUrl } from './edge-proxy';
 export { SitecoreClient, Page, PageOptions, SitemapXmlOptions } from './sitecore-client';
 export { SitecoreClientInit } from './models';
 export { createGraphQLClientFactory, GraphQLClientOptions } from './utils';